Parser is coming along
At this stage, we can now parse type and method declarations, and a single
statement form (assert e).  This is enough to now push ahead and put in place a
REPL and then a type checker.  For example, "assert 1;" should fail to type
check but will pass.

In order to get decent error messagse out of the type checker, however, requires
some quite drastic changes.  In particular, we need to support annotations on our
syntatic elements.  This allows us to include source information, and also type
information.