Working on reinstating type checking
There are a few questions outstanding about type checking.  Firstly,
various methods return instances of Type, not instances of Node.
Therefore, we must have those instances allocated in the heap.  This
makes sense since the source map will require that as well.

The other issue is that the check_subtype method is currently operating
on two instances of Type.  But, in many cases, we're not comparing two
things from the heap, but one thing we "know" and one thing from the
heap.