Update Type Checker
Type checking is now working, though it has raised some interesting
design issues.  In particular, every expression is given a type and said
type must be in the AST.  Therefore, said type is added to the AST which
means the type checker must take "&mut ast".  However, this does impact
the structure of the traversal in the type checker.  To resolve this,
for now, I have introduced some cloning which is annoying.  There are at
least two reasonable ways forward:

(1) All non-copy types should be loaded into the AST (e.g. name: String,
Vec<Parameter>, etc).

(2) Ensure the type checker doesn't have to mutate the ast, other than
to update the type map.  This is only possible if the type checker
doesn't have to make "types out of thin air".  This may hold (in this
case) if all types must already be given in the source (except for a
small number of primitive types such as boolean, etc).