Add Node::Name
This makes a special node kind for representing strings or identifiers.  This
reduces the need to clone data when traversing the AST.  An outstanding issue is
that the parser does not fold names together and, hence, environment lookups
fail resulting in "variable not found" errors.

In addition, this commit turns Type into a tuple struct.